requirements.txt
arrow @ file:///home/conda/feedstock_root/build_artifacts/arrow_1610483722687/work
binaryornot==0.4.4
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1606255081583/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093490430/work
click==7.1.2
cookiecutter==1.7.2
cryptography @ file:///tmp/build/80754af9/cryptography_1607635341180/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
Jinja2==2.11.2
jinja2-time==0.2.0
kaggle==1.5.10
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1602267312178/work
poyo==0.5.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291447907/work
python-dateutil==2.8.1
python-slugify @ file:///home/conda/feedstock_root/build_artifacts/python-slugify_1593573453419/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
text-unidecode==1.3
tqdm==4.56.0
Unidecode @ file:///home/conda/feedstock_root/build_artifacts/unidecode_1608471621084/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1605225765842/work
whichcraft==0.6.1

pandas~=1.2.0
tokenizers~=0.9.4
python-dotenv~=0.15.0
setuptools~=49.6.0
torch~=1.7.1
transformers~=4.2.1
wandb~=0.10.14
numpy~=1.19.5
sklearn~=0.0
scikit-learn~=0.24.0